requirements-test.txt
pytest
pytest-cov
mockredispy
